Add check to disallow deletion of enviromnents used by applications
Controller checks that the environment is not used in any applications. If it is, pass warning message.

Added test.

+++ laravel/app/Http/Controllers/EnvironmentsController.php
@@ -105,7 +105,11 @@ class EnvironmentsController extends Controller
      */
     public function destroy(Environments $environment)
     {
-        $environment->delete();
+        if (!$environment->applications->count()) {
+            $environment->delete();
+        } else {
+            session()->flash('warning', "Can't delete environment being used by applications");
+        }
 
         return redirect('/environments');
     }
